提交

+++ src/pioneer.h
@@ -7,15 +7,15 @@
 
 #define INDEXS_SIZE 1024
 
-int64_t HistLen = 0;
-uint64_t * Time;
+int HistLen = 0;
+unsigned long * Time;
 double * Open;
 double * High;
 double * Low;
 double * Close;
 double * Volume;
 double * Indexs[INDEXS_SIZE];
-int64_t StablePoint = 0;
+int StablePoint = 0;
 // 初始资金
 double init_funds = 100.0;
 // 实时资金
@@ -29,6 +29,6 @@ double funds_buy = 0.0;
 // 历史最大资金
 double funds_max = 0.0;
 // 盈利交易计数
-int64_t win_count = 0;
+int win_count = 0;
 // 亏损交易计数
-int64_t loss_count = 0;
+int loss_count = 0;
